Move Views de wagtail_hooks para views
