+++ proc/wagtail_hooks.py
@@ -1,6 +1,3 @@
-from django.contrib import messages
-from django.http import HttpResponseRedirect
-from django.shortcuts import redirect
 from django.urls import include, path
 from django.utils.translation import gettext as _
 from wagtail import hooks
@@ -9,31 +6,31 @@ from wagtail.contrib.modeladmin.options import (
     ModelAdminGroup,
     modeladmin_register,
 )
-from wagtail.contrib.modeladmin.views import CreateView, EditView, InspectView
+from wagtail.contrib.modeladmin.views import InspectView
 
 from config.menu import get_menu_order
 from htmlxml.models import HTMLXML
 from package.models import SPSPkg
 
 from .models import ArticleProc, IssueProc, JournalProc, ProcReport
+from proc.views import ProcCreateView, ProcEditView, CoreCreateView
+
+# class ProcCreateView(CreateView):
+#     def form_valid(self, form):
+#         self.object = form.save_all(self.request.user)
+#         return HttpResponseRedirect(self.get_success_url())
 
 
-class ProcCreateView(CreateView):
-    def form_valid(self, form):
-        self.object = form.save_all(self.request.user)
-        return HttpResponseRedirect(self.get_success_url())
+# class ProcEditView(EditView):
+#     def form_valid(self, form):
+#         self.object = form.save_all(self.request.user)
+#         return HttpResponseRedirect(self.get_success_url())
 
 
-class ProcEditView(EditView):
-    def form_valid(self, form):
-        self.object = form.save_all(self.request.user)
-        return HttpResponseRedirect(self.get_success_url())
-
-
-class CoreCreateView(CreateView):
-    def form_valid(self, form):
-        self.object = form.save_all(self.request.user)
-        return HttpResponseRedirect(self.get_success_url())
+# class CoreCreateView(CreateView):
+#     def form_valid(self, form):
+#         self.object = form.save_all(self.request.user)
+#         return HttpResponseRedirect(self.get_success_url())
 
 
 class JournalProcModelAdmin(ModelAdmin):
@@ -43,6 +40,7 @@ class JournalProcModelAdmin(ModelAdmin):
     menu_order = 200
     add_to_settings_menu = False
     exclude_from_explorer = False
+    create_view_class = ProcCreateView
     edit_view_class = ProcEditView
 
     list_display = (
